완성!!!!

+++ MRCNN/utils.py
@@ -863,7 +863,7 @@ def resize(image, output_shape, order=1, mode='constant', cval=0, clip=True,
     version. And it provides a central place to control resizing defaults.
     """
     return skimage.transform.resize(
-        image.astype(np.uint8), output_shape,
+        image.astype(np.float32), output_shape,
         order=order, mode=mode, cval=cval, clip=clip,
         preserve_range=preserve_range, anti_aliasing=anti_aliasing,
         anti_aliasing_sigma=anti_aliasing_sigma)
